+++ Share2PleyApp/src/com/share2pley/share2pleyapp/SetPagerFragment.java
@@ -1,24 +1,22 @@
 package com.share2pley.share2pleyapp;
 
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.util.ArrayList;
 
+import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.Intent;
 import android.content.res.AssetManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.share2pley.share2pleyapp.Model.DatabaseSetHelper;
 import com.share2pley.share2pleyapp.Model.SetLab;
@@ -106,46 +104,28 @@ public class SetPagerFragment extends Fragment {
 		instructionsButton.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View v) {
-				copyReadAssets();
+				try {
+					startPdf(mPageNumber);
+				} catch (ActivityNotFoundException e) {
+					Toast.makeText(getActivity(), "No PDF reader installed.",
+							100).show();
+				}
 			}
 		});
 	}
 
-	private void copyReadAssets() {
+	private void startPdf(int pageNumber) {
 		AssetManager assetManager = getActivity().getAssets();
-
-		InputStream in = null;
-		OutputStream out = null;
-		File file = new File(getActivity().getFilesDir(), "speelhuis.pdf");
 		try {
-			in = assetManager.open("speelhuis.pdf");
-			out = getActivity().openFileOutput(file.getName(),
-					Context.MODE_WORLD_READABLE);
-
-			copyFile(in, out);
-			in.close();
-			in = null;
-			out.flush();
-			out.close();
-			out = null;
-		} catch (Exception e) {
-			Log.e("tag", e.getMessage());
-		}
-
-		Intent intent = new Intent(Intent.ACTION_VIEW);
-		intent.setDataAndType(
-				Uri.parse("file://" + getActivity().getFilesDir()
-						+ "/speelhuis.pdf"), "application/pdf");
-
-		startActivity(intent);
-	}
-
-	private void copyFile(InputStream in, OutputStream out) throws IOException {
-		byte[] buffer = new byte[1024];
-		int read;
-		while ((read = in.read(buffer)) != -1) {
-			out.write(buffer, 0, read);
+			String[] files = assetManager.list("");
+			Intent intent = new Intent(Intent.ACTION_VIEW);
+			String[] array = getResources().getStringArray(R.array.assetsArray);
+			intent.setDataAndType(
+					Uri.parse("file://" + getActivity().getFilesDir()
+							+ Uri.parse(array[pageNumber])), "application/pdf");
+			startActivity(intent);
+		} catch (IOException e) {
+			Toast.makeText(getActivity(), "It is not working", 100);
 		}
 	}
-
 }
